refactor(spec): split style data-checks into finite and numeric modules
Extract shape/linetype finite checks and size/linewidth/alpha numeric
checks from validate-data-checks-style.ts. Keep a thin barrel so the
data-checks orchestrator import path stays stable.
