+++ packages/spec/src/validate-data-checks-position.ts
@@ -6,7 +6,7 @@
  *  - checkPositionScaleDataCompatibility — post-layer scale/type vs field evidence
  *
  * Shared temporal memoization: validate-data-checks-temporal.ts.
- * Color: validate-data-checks-color.ts. Style: validate-data-checks-style.ts.
+ * Color: validate-data-checks-color.ts. Style: validate-data-checks-style.ts (finite + numeric).
  * Orchestrator: validate-data-checks.ts.
  */
 import type { SpecError } from "./errors.js";
