+++ packages/spec/src/validate-data-checks.ts
@@ -7,7 +7,7 @@
  *  - validate-data-checks-temporal.ts — shared temporalDecisionForField memoization
  *  - validate-data-checks-position.ts — pre-evidence temporal axis config + x/y scale types
  *  - validate-data-checks-color.ts — color/fill scale types + manual domain/range
- *  - validate-data-checks-style.ts — shape/linetype + size/linewidth/alpha scale types
+ *  - validate-data-checks-style.ts — shape/linetype + size/linewidth/alpha (finite + numeric modules)
  *
  * Style and color scale checks take the same per-use `evidenceForUse` path as
  * position (#609 / #844): multi-table layers that share a field name keep their
