+++ packages/spec/src/validate-data.ts
@@ -11,7 +11,7 @@
  *  - validate-data-checks-temporal.ts — shared temporalDecisionForField memoization
  *  - validate-data-checks-position.ts — pre-evidence temporal config + x/y scale types
  *  - validate-data-checks-color.ts — color/fill scale types + manual domain/range
- *  - validate-data-checks-style.ts — shape/linetype + size/linewidth/alpha scale types
+ *  - validate-data-checks-style.ts — shape/linetype + size/linewidth/alpha (finite + numeric modules)
  *
  * Style and color scale checks use per-layer evidenceForUse like position
  * (#609 / #844) so multi-table same-name fields keep independent type views.
